+++ calour/util.py
@@ -27,6 +27,8 @@ import inspect
 import configparser
 from pkg_resources import resource_filename
 from collections import Iterable
+from numbers import Number
+
 import scipy
 
 
@@ -318,3 +320,35 @@ def _to_list(x):
     if isinstance(x, Iterable):
         return x
     return [x]
+
+
+def _argsort(values):
+    '''Do an argsort separately on numeric and non-numeric values, then combine (numeric first)
+
+    Used to overcome the problem when using numpy.argsort on a pandas series values with missing values
+
+    Parameters
+    ----------
+    values : iterable
+        the values to sort
+
+    Returns
+    -------
+    tuple of ints
+        the positions of the sorted values
+    '''
+    # convert all numbers to float otherwise int will be sorted different place
+    values = [float(x) if isinstance(x, Number) else x for x in values]
+    v = zip(values, list(range(len(values))))
+    res = sorted(v, key=lambda x: (str(type(x[0])), x))
+    return [x[1] for x in res]
+    # nums = [x for x in v if isinstance(x[0], Number)]
+    # v = zip(values, list(range(len(values))))
+    # strs = [x for x in v if not isinstance(x[0], Number)]
+    # if len(nums) > 0 and len(strs) > 0:
+    #     logger.warn('Values contain mixed numbers and strings. Sorted separately.')
+    # nums = sorted(nums, key=lambda x: x[0])
+    # strs = sorted(strs, key=lambda x: str(x[0]))
+    # allvals = [x[1] for x in nums]
+    # allvals.extend([x[1] for x in strs])
+    # return allvals
